pyproject.toml
[project]
name = "tagger"
version = "0.1.0"
description = "DSPy-based hierarchical text classifier"
requires-python = ">=3.12"
dependencies = [
    "dspy-ai>=2.5.0",
    "pydantic>=2.0.0",
    "gepa>=0.0.26",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
